.dagger/workspace/pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["dagger-io", "PyGithub>=2.1.1"]

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[project.entry-points."dagger.mod"]
main_object = "workspace:Workspace"
